add fineHostInstance
